Refactor with new package suggestion
